move store functions and settings into own file
